backend/requirements.txt
# backend/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9          # multipart/form-data uploads

# ── ML / Image ────────────────────────────────────────────────────────────
keras==3.13.2
tensorflow==2.17.0               # Keras 3 default backend (swap for jax/torch)
numpy==1.26.4
Pillow==10.4.0
piexif==1.1.3                    # EXIF metadata stripping (HIPAA PHI removal)
opencv-python-headless==4.10.0.84

# ── Database ──────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2

# ── Auth & Security ───────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── AWS ───────────────────────────────────────────────────────────────────
boto3==1.34.144

# ── Config ────────────────────────────────────────────────────────────────
pydantic-settings==2.3.4

# ── Dev / Testing ─────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.8
httpx==0.27.0
